Moved Lcc/Vsc round trip tests in HvdcRoundTripTest.cpp and move shunt round trip tests in ShuntCompensatorRoundTripTest.cpp (#247)
